requirements.txt
flask==3.0.0
flask-cors==4.0.0
twilio==8.10.0
python-dotenv==1.0.0
apscheduler==3.10.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
gunicorn==21.2.0
